+++ request.h
@@ -15,26 +15,26 @@ public:
          m_type(t), m_seg(s){}
     Segment & getSegment(){return m_seg;}
     RequestType getType(){return m_type;}
-    float getSendTime(){return m_send_time;}
+    double getSendTime(){return m_send_time;}
 
     bool m_is_media_buffering;
 
 private:
     RequestType m_type;
     Segment m_seg;
-    float m_received_time = 0;
-    float m_send_time = 0;
+    double m_received_time = 0;
+    double m_send_time = 0;
 
 };
 
 class Downloader{
 private:
-    static int m_number;
+    static unsigned int m_number;
     std::vector<unsigned int> & B;
     std::vector<Request> R;
 public:
     Downloader(std::vector<unsigned int> & bitrates);
-    float get(Request & r, float at_time);
+    double get(Request & r, double at_time);
     std::vector<Request> & getRequests(){return R;}
 };
